Various improvements and fixes. Notably switched from klepto to sqlitedict + dill.
One caveat is that (this is true for the old klepto implementation as well) the persistent archives provided by AbstractProcessor (and the ones used by ReactiveBuilder by extension) are cached at class level --- in other words, all instances of the same builder/processor class share the same persistence archive.  This can be a surprising behaviour.